Bug fix and cryptographic 'strengthening'
Fixed a bug where if the RNG returned < 1000, it would fail to continue generating numbers. Bit of a hacky solution, but is a solution nonetheless. The rng could do with a rewrite to be honest, but that's another issue for another commit.

Strengthened the crypto by reversing the plaintext before encrypting, and then reversing the ciphertext after encrypting. This requires that an attacker must get the entire message to potentially launch an attack (at least... I think). Also, partial strings won't have the same ciphertext.