Added permutation functions, refactoring
Added two functions (permute and unpermute) that will perform a permutation of the input text using the key as an index to swap characters around (and reverse permutation on decrypt).

Due to how the permutation works, this required specifying an encrypt and decrypt operation (using -d and -e flags). Also refactored some code, generalised input and output, rather than plaintext and ciphertext.